lab 5

+++ 220041212_task2_lab4.cpp
@@ -9,7 +9,7 @@ private:
     string accountholdername;
     string accounttype;
     float currentbalance;
-    const float minimumvalue = 100;
+    const int  minimumvalue = 100;
 
 public:
     bankaccount(int accountnumber, string name, string accounttype, float currentbalance)
